Implement basic movement paradigm
The basic idea for move checking is this: The Board is in charge of
checking that squares are not empty and that a square contains a piece.
Eventually I think that checking for checkmate, stalemate, etc. will be
a Board responsibility also. But as far as the Board is concerned, as
long as a piece doesn't move off the board then it's a valid move.

The individual pieces, on the other hand, are responsible for checking
whether a move is valid or not *for that particular piece*. A pawn will
check whether it's only moving one space (or two spaces in the
beginning) at a time, it will check that it's not running into a another
 piece, that it's staying in the same column (unless it's attacking),
 etc. It will use public methods in Board (like isOccupied,
 isVerticalMove, getMoveLength, etc.) to accomplish this. There are a
 number of benefits here:

 - First and foremost, it separates responsibility between the two
 classes (Board and whichever piece is moving).
 - Second, it ensures proper permissions. The only things that board
 exposes are the public methods that the Board checks itself. We don't
 have to pass around the container (map) of squares, because the Board
 is checking them itself.
 - Third, it allows us to reuse a lot of code. Many pieces (rook, queen,
  pawn, etc.) will need to know, for example, how far they can move
  vertically before they hit another piece. We can maintain one function
  that all of those pieces can take advantage of.
- Finally, it leaves Board management (i.e., moving pieces) up to the
Board. All the piece does is report whether a given move is valid for
that particular piece.

In addition to the above changes, I also made a quick switch from using
to std::pair<int,int>, such that the algebraic location a8 is
std::pair{0,0}. This has a number of benefits, namely easier square
instantiation and move checking. Made several updates to take advantage
of this new functionality, like simplifying printing of the board. Will
implement a function to convert between algebraic and pair/int notation
later. This follows Daniel's example, except that pair has the added
benefit of not requiring a function call to access its contents, they're
 exposed as member variables.