Implement isClearPath
- Add function isClearPath
- Add unit tests covering different inputs and outputs to isClearPath

The ability to check whether a path from one location to another is a
critical requirement of all pieces except the knight (which has its own
move characteristics) and the king (which only ever moves one step at a
time). Nesting this function in the Board class grants this same
functionality to all pieces.