Fix and improve movement logic (#11)
- Changed specifier for getPiece()
- Create extra functions for occupied by same or different color

Noticed a problem with pawn movement, namely that a pawn was able to
move one square and then move again two squares. Problem was that the
moved member variable for pawn wasn't being updated. So, modified
getPiece (removing const specifier) so that moved could be modified.

Moving a piece to a location occupied by a piece of the same color is
never allowed, so that check now takes place at the Board level before
calling an individual piece's isValidMove function. For specific piece
types, you can either call isOccupied if the color doesn't matter (as in
a pawn that has a piece in front of it) or isOccupiedDifferentColor to
check for an attack (for example).