Fix error in diagonal clear path
Tony previously found and fixed an error regarding checking whether
diagonal paths are clear. This commit extends that fix to the other
diagonal direction and updates unit tests to match. Also renames i and j
to row and col, respectively.

Also moved the testing cases to CHECK rather than REQUIRE. CHECK will
allow other tests to occur, whereas if a REQUIRE fails then computation
is stopped and no other tests are run.

Finally, whether or not a path is clear has nothing to do with whether the 
destination is a friendly or non-friendly piece. Indeed, it doesn't check the 
destination at all. So I've removed that block in isPathClear().