Implement basic gameplay
- Add printing function which prints the axes in algebraic coordinates
- Add unit tests for said function
- Move code in Game::run() to Game::testRun()
- Refactor Game::run() as a command-line chess game

When I say the gameplay is basic, I mean basic. There's no turn
checking, no game state checking. Nada. As Tony (I think) mentioned,
these sorts of checks will eventually go in the Game class, since
stalemate is a state of a Game more than it is a state of a Board.

I will expand on this as time provides, but I wanted to have something
working and running by Friday, however minimal.