Implement castling
- Remove castling check from King
- Remove isValidCastle() from Board
- Add setRook() to Board
- Add castle() to Game
- Add castle functionality to game loop
- Add two unit tests (black queenside and white kingside)

The basic idea here is that castling is implemented at the Game level
since it relies on the game state of check. The castling movement is
broken up into two steps. After each step we check whether the player is
in check. If so, we back up accordingly. Assuming this is a valid
castle, we finish by placing the rook in the space that the king jumped
over.